Apply changes to etcdv2 code after package move
