+++ subnet/etcdv2/local_manager.go
@@ -380,3 +380,21 @@ func isSubnetConfigCompat(config *Config, sn ip.IP4Net) bool {
 func (m *LocalManager) Name() string {
 	return fmt.Sprintf("Etcd Local Manager with Previous Subnet: %s", m.previousSubnet.String())
 }
+
+func (m *LocalManager) CreateBackendData(ctx context.Context, network, data string) error {
+	err := m.registry.createBackendData(ctx, network, data)
+
+	if err == nil {
+		return nil
+	}
+
+	if etcdErr, ok := err.(etcd.Error); ok && etcdErr.Code == etcd.ErrorCodeNodeExist {
+		return nil
+	}
+
+	return err
+}
+
+func (m *LocalManager) GetBackendData(ctx context.Context, network string) (string, error) {
+	return m.registry.getBackendData(ctx, network)
+}
