+++ subnet/etcdv2/registry.go
@@ -45,6 +45,8 @@ type Registry interface {
 	deleteSubnet(ctx context.Context, sn ip.IP4Net) error
 	watchSubnets(ctx context.Context, since uint64) (Event, uint64, error)
 	watchSubnet(ctx context.Context, since uint64, sn ip.IP4Net) (Event, uint64, error)
+	createBackendData(ctx context.Context, network, data string) error
+	getBackendData(ctx context.Context, network string) (string, error)
 }
 
 type EtcdConfig struct {
@@ -296,6 +298,27 @@ func parseSubnetWatchResponse(resp *etcd.Response) (Event, error) {
 		}
 		return evt, nil
 	}
+}
+
+func (esr *etcdSubnetRegistry) createBackendData(ctx context.Context, network, data string) error {
+	key := path.Join(esr.etcdCfg.Prefix, network, "backend-data")
+	_, err := esr.client().Create(ctx, key, data)
+	if err != nil {
+		return err
+	}
+
+	return nil
+}
+
+func (esr *etcdSubnetRegistry) getBackendData(ctx context.Context, network string) (string, error) {
+	key := path.Join(esr.etcdCfg.Prefix, network, "backend-data")
+	etcdResponse, err := esr.client().Get(ctx, key, nil)
+
+	if err != nil {
+		return "", err
+	}
+
+	return etcdResponse.Node.Value, nil
 }
 
 func nodeToLease(node *etcd.Node) (*Lease, error) {
